+++ lib/onlyoffice_tcm_helper.rb
@@ -61,7 +61,12 @@ module OnlyofficeTcmHelper
     # @return [Symbol] is a symbolic name of test status
     def get_status_and_comment(example)
       @status = if result_is_failed?(example)
-                  @comment = "\n#{example.exception.to_s.gsub('got:', "got:\n").gsub('expected:', "expected:\n").gsub('to return ', "to return:\n")}\nIn line:\n#{RspecHelper.find_failed_line(example)}"
+                  comment = "\n#{example.exception.to_s
+                                        .gsub('got:', "got:\n")
+                                        .gsub('expected:', "expected:\n")
+                                        .gsub('to return ', "to return:\n")}\n"\
+                            "In line:\n#{RspecHelper.find_failed_line(example)}"
+                  @comment = comment
                   :failed
                 elsif example.pending?
                   @comment = example.execution_result.pending_message