+++ lib/onlyoffice_tcm_helper/models/pseudo_example_failed/mock_failed_exception.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # Mock class to check exception
 class MockFailedException
   def to_s
@@ -8,7 +10,6 @@ class MockFailedException
   def backtrace
     ['trace_without_line',
      "#{Dir.pwd}/spec/onlyoffice_tcm_helper_spec.rb:3:in `get_status_and_comment",
-     "#{Dir.pwd}/spec/onlyoffice_tcm_helper_spec.rb:4:in `parse"
-     ]
+     "#{Dir.pwd}/spec/onlyoffice_tcm_helper_spec.rb:4:in `parse"]
   end
 end