+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add `markdownlint` support and check in CI
 * Add `simplecov` support to check coverage
 * Add coverage check in CI
+* Add `rubocop` check to CI
 
 ### Changes